setting: Always remove Left-handed mode setting for non-phones
Right now, this option is only removed if you have a navigation bar
enabled and you aren't a phone.  This leads to UI wonkiness on
devices that allow the navigation bar to be toggled on/off:

* Enable the navigation bar.
* Navigate elsewhere and back to Buttons (force it to reload)
* Settings >> Buttons now contains a (grayed out) Left-handed mode.
* Enable the navigation bar.

Now Left-handed mode is available as an option even though it
shouldn't be.

Change-Id: Ifb3cdd18a014edffaf8bc53c46fc3c2f7a681370

Conflicts:
	src/com/android/settings/cyanogenmod/ButtonSettings.java